Move old ES utilities code to tests
Because some code is still used in tests.
