+++ libraries/utilities/include/graphene/utilities/elasticsearch.hpp
@@ -27,7 +27,7 @@
 #include <vector>
 
 #include <curl/curl.h>
-#include <fc/time.hpp>
+
 #include <fc/variant_object.hpp>
 
 namespace graphene { namespace utilities {
@@ -111,36 +111,7 @@ private:
    curl_wrapper curl;
 };
 
-   class ES {
-      public:
-         CURL *curl;
-         std::vector <std::string> bulk_lines;
-         std::string elasticsearch_url;
-         std::string index_prefix;
-         std::string auth;
-         std::string endpoint;
-         std::string query;
-   };
-   class CurlRequest {
-      public:
-         CURL *handler;
-         std::string url;
-         std::string type;
-         std::string auth;
-         std::string query;
-   };
-
-   bool SendBulk(ES&& es);
-   bool checkES(ES& es);
-   std::string getESVersion(ES& es);
-   void checkESVersion7OrAbove(ES& es, bool& result) noexcept;
-   std::string simpleQuery(ES& es);
-   bool deleteAll(ES& es);
-   std::string getEndPoint(ES& es);
-
-   std::string doCurl(CurlRequest& curl);
-
-   std::vector<std::string> createBulk(const fc::mutable_variant_object& bulk_header, std::string&& data);
+std::vector<std::string> createBulk(const fc::mutable_variant_object& bulk_header, std::string&& data);
 
 struct es_data_adaptor
 {
